Move constants out of util.py.
