+++ gslib/boto_translation.py
@@ -89,14 +89,14 @@ from gslib.translation_helper import LifecycleTranslation
 from gslib.translation_helper import REMOVE_CORS_CONFIG
 from gslib.translation_helper import S3MarkerAclFromObjectMetadata
 from gslib.util import AddAcceptEncodingGzipIfNeeded
-from gslib.util import DEFAULT_FILE_BUFFER_SIZE
 from gslib.util import S3_DELETE_MARKER_GUID
 from gslib.util import UnaryDictToXml
-from gslib.util import XML_PROGRESS_CALLBACKS
 from gslib.utils.boto_util import ConfigureNoOpAuthIfNeeded
 from gslib.utils.boto_util import GetMaxRetryDelay
 from gslib.utils.boto_util import GetNumRetries
+from gslib.utils.constants import DEFAULT_FILE_BUFFER_SIZE
 from gslib.utils.constants import UTF8
+from gslib.utils.constants import XML_PROGRESS_CALLBACKS
 from gslib.utils.unit_util import TWO_MIB
 
 TRANSLATABLE_BOTO_EXCEPTIONS = (boto.exception.BotoServerError,
