+++ gslib/copy_helper.py
@@ -125,16 +125,16 @@ from gslib.translation_helper import PreconditionsFromHeaders
 from gslib.translation_helper import S3MarkerAclFromObjectMetadata
 from gslib.util import CheckMultiprocessingAvailableAndInit
 from gslib.util import CreateLock
-from gslib.util import DEFAULT_FILE_BUFFER_SIZE
 from gslib.util import GetCloudApiInstance
 from gslib.util import IsCloudSubdirPlaceholder
-from gslib.util import MIN_SIZE_COMPUTE_LOGGING
 from gslib.util import ObjectIsGzipEncoded
 from gslib.utils.boto_util import GetJsonResumableChunkSize
 from gslib.utils.boto_util import GetMaxRetryDelay
 from gslib.utils.boto_util import GetNumRetries
 from gslib.utils.boto_util import ResumableThreshold
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.constants import DEFAULT_FILE_BUFFER_SIZE
+from gslib.utils.constants import MIN_SIZE_COMPUTE_LOGGING
 from gslib.utils.constants import UTF8
 from gslib.utils.system_util import CheckFreeSpace
 from gslib.utils.system_util import GetFileSize
