+++ gslib/gcs_json_media.py
@@ -30,9 +30,9 @@ from apitools.base.py import exceptions as apitools_exceptions
 from gslib.cloud_api import BadRequestException
 from gslib.lazy_wrapper import LazyWrapper
 from gslib.progress_callback import ProgressCallbackWithTimeout
-from gslib.util import DEBUGLEVEL_DUMP_REQUESTS
-from gslib.util import TRANSFER_BUFFER_SIZE
+from gslib.utils.constants import DEBUGLEVEL_DUMP_REQUESTS
 from gslib.utils.constants import SSL_TIMEOUT_SEC
+from gslib.utils.constants import TRANSFER_BUFFER_SIZE
 import httplib2
 from httplib2 import parse_uri
