+++ gslib/tests/test_metrics.py
@@ -48,7 +48,7 @@ from gslib.third_party.storage_apitools import storage_v1_messages as apitools_m
 from gslib.thread_message import FileMessage
 from gslib.thread_message import RetryableErrorMessage
 from gslib.util import LogAndHandleRetries
-from gslib.util import START_CALLBACK_PER_BYTES
+from gslib.utils.constants import START_CALLBACK_PER_BYTES
 from gslib.utils.system_util import IS_LINUX
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.utils.unit_util import ONE_KIB
