+++ gslib/tests/test_ui.py
@@ -66,8 +66,8 @@ from gslib.ui_controller import MetadataManager
 from gslib.ui_controller import UIController
 from gslib.ui_controller import UIThread
 from gslib.util import Retry
-from gslib.util import START_CALLBACK_PER_BYTES
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.constants import START_CALLBACK_PER_BYTES
 from gslib.utils.constants import UTF8
 from gslib.utils.unit_util import HumanReadableWithDecimalPlaces
 from gslib.utils.unit_util import MakeHumanReadable
